# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TEST API SMART LOGISTICS",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "api",
            "description": "Local server"
        }
    ],
    "paths": {
        "/slots/availability": {
            "get": {
                "tags": [
                    "Слоты"
                ],
                "summary": "Получение доступных слотов",
                "operationId": "getSlotsAvailability",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Сортировка (например: id_asc, id_desc, remaining_asc, remaining_desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "id_asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список доступных слотов с пагинацией",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "slot_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "capacity": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "remaining": {
                                                        "type": "integer",
                                                        "example": 6
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/slots/{id}/hold": {
            "post": {
                "tags": [
                    "Холды"
                ],
                "summary": "Создание холда (бронирования места)",
                "operationId": "f74dc621153c7063a2a638f8d81442ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Холд успешно создан"
                    },
                    "409": {
                        "description": "Конфликт: нет свободных мест"
                    },
                    "400": {
                        "description": "Некорректный UUID в заголовке"
                    }
                }
            }
        },
        "/holds/{id}/confirm": {
            "post": {
                "tags": [
                    "Холды"
                ],
                "summary": "Подтверждение холда",
                "operationId": "confirmHold",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Холд успешно подтвержден"
                    },
                    "409": {
                        "description": "Конфликт: нет свободных мест или холд не в статусе held"
                    },
                    "404": {
                        "description": "Холд не найден"
                    }
                }
            }
        },
        "/holds/{id}": {
            "delete": {
                "tags": [
                    "Холды"
                ],
                "summary": "Отмена холда",
                "operationId": "cancelHold",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Холд успешно отменен"
                    },
                    "404": {
                        "description": "Холд не найден"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Слоты",
            "description": "Слоты"
        },
        {
            "name": "Холды",
            "description": "Холды"
        }
    ]
}